Fixing PR as requested (2)
