+++ lib/manager/nuget/__fixtures__/msbuild-sdk-files/global.json
