+++ lib/manager/nuget/__fixtures__/msbuild-sdk-files/invalid-json/global.json
