+++ lib/manager/nuget/__fixtures__/msbuild-sdk-files/not-nuget/global.json
